Centralize all exceptions
Move all exceptions to the gafaelfawr.exceptions module, which
lays the groundwork for some additional consistency, coherent
exception hierarchies, and helper functions.
