+++ src/gafaelfawr/providers/oidc.py
@@ -7,10 +7,10 @@ from urllib.parse import urlencode
 
 import jwt
 
-from gafaelfawr.providers.base import Provider, ProviderException
+from gafaelfawr.exceptions import OIDCException, VerifyTokenException
+from gafaelfawr.providers.base import Provider
 from gafaelfawr.session import Session, SessionHandle
 from gafaelfawr.tokens import Token
-from gafaelfawr.verify import VerifyTokenException
 
 if TYPE_CHECKING:
     from aiohttp import ClientSession
@@ -21,11 +21,7 @@ if TYPE_CHECKING:
     from gafaelfawr.session import SessionStore
     from gafaelfawr.verify import TokenVerifier
 
-__all__ = ["OIDCException", "OIDCProvider"]
-
-
-class OIDCException(ProviderException):
-    """The OpenID Connect provider returned an error from an API call."""
+__all__ = ["OIDCProvider"]
 
 
 class OIDCProvider(Provider):
@@ -114,7 +110,7 @@ class OIDCProvider(Provider):
             provider.
         jwt.exceptions.InvalidTokenError
             The token returned by the OpenID Connect provider was invalid.
-        OIDCException
+        gafaelfawr.exceptions.OIDCException
             The OpenID Connect provider responded with an error to a request.
         """
         data = {
